Set of fixes
